Version 3 initial refactor
