+++ src/main.rs
@@ -1,4 +1,4 @@
-// Copyright (c) 2016 mussh developers
+// Copyright (c) 2016, 2018 mussh developers
 //
 // Licensed under the Apache License, Version 2.0
 // <LICENSE-APACHE or http://www.apache.org/licenses/LICENSE-2.0> or the MIT
@@ -7,42 +7,66 @@
 // modified, or distributed except according to those terms.
 
 //! mussh - SSH Multiplexing
-#![cfg_attr(feature = "cargo-clippy", allow(unseparated_literal_suffix, use_self))]
-#![recursion_limit = "128"]
-#![deny(missing_docs)]
-#[macro_use]
-extern crate error_chain;
-#[macro_use]
-extern crate serde_derive;
-#[macro_use]
-extern crate slog;
+#![feature(crate_visibility_modifier, try_from)]
+#![deny(
+    clippy::all,
+    clippy::pedantic,
+    macro_use_extern_crate,
+    missing_copy_implementations,
+    missing_debug_implementations,
+    missing_docs,
+    trivial_casts,
+    trivial_numeric_casts,
+    unused
+)]
+#![warn(
+    absolute_paths_not_starting_with_crate,
+    anonymous_parameters,
+    bare_trait_objects,
+    box_pointers,
+    elided_lifetimes_in_paths,
+    ellipsis_inclusive_range_patterns,
+    keyword_idents,
+    question_mark_macro_sep,
+    single_use_lifetimes,
+    unreachable_pub,
+    unsafe_code,
+    unused_extern_crates,
+    unused_import_braces,
+    unused_labels,
+    unused_lifetimes,
+    unused_qualifications,
+    unused_results,
+    variant_size_differences
+)]
+#![allow(clippy::stutter)]
+#![doc(html_root_url = "https://docs.rs/mussh/3.0.0")]
 
-extern crate chrono;
-extern crate clap;
-extern crate dirs;
-extern crate slog_async;
-extern crate slog_term;
-extern crate ssh2;
-extern crate sys_info;
-extern crate term;
-extern crate toml;
-
-mod cmd;
-mod config;
-mod error;
 mod run;
-mod util;
+mod subcmd;
 
-use std::io::{self, Write};
+use clap::ErrorKind;
 use std::process;
 
 /// mussh entry point
 fn main() {
     match run::run() {
-        Ok(i) => process::exit(i),
-        Err(e) => {
-            writeln!(io::stderr(), "{}", e).expect("badness");
-            process::exit(1)
+        Ok(_) => process::exit(0),
+        Err(error) => {
+            let cause = error.as_fail();
+
+            if let Some(err) = cause.downcast_ref::<clap::Error>() {
+                let kind = err.kind;
+                eprintln!("{}", err.message);
+                match kind {
+                    ErrorKind::HelpDisplayed | ErrorKind::VersionDisplayed => process::exit(0),
+                    _ => process::exit(1),
+                }
+            } else {
+                eprintln!("{}", error.as_fail());
+                eprintln!("{}", error.backtrace());
+                process::exit(1);
+            }
         }
     }
 }
